Module libre_pvz_animation::curve

Variable curves in animation clips.

Modules

animatable

Types that may be used in curves.

blend

Blending two different animation segments.

builder

Builder API for Curves.

concrete

Concrete curves.

Structs

CurveBinding

Bind a contiguous range of curves (on the same component) to some entity.

CurveBindingInfo

Information about a curve binding.

CurveDescriptor

Descriptor for the curves. Different curves for the same component share the same descriptor.

Segment

A segment in a curve.

UnwrapAnyComponent

Assert the target component type for some AnyComponent.

Traits

AnyComponent

Bevy tracks changes in mutable references like Mut. We use this interface to avoid unnecessarily marking the target component as changed.

AnyCurve

Type-erased Curve. Prefer Curve whenever possible.

Curve

Animation curve.

TypedCurve

Animation curve with known single variable type.