Enum libre_pvz_animation::curve::blend::BlendMethod

pub enum BlendMethod {
    Linear,
    Smooth,
    SmoothTanh(f32),
}

The blend function for use. All functions below transitions from (0, 0) to (1, 1).

Below is a plot of the three supported blend methods:

Linear: the orange line

Smooth: the green line

SmoothTanh (with α=1.5): the pink line

Variants

Linear

A simple linear transition function:
ϕ ( x ) = x ( 0 ≤ x ≤ 1 )

Smooth

A smooth transition function using the exponential function:
ϕ ( x ) = exp ⁡ ( − 1 ⁄ x ) exp ⁡ ( − 1 ⁄ x ) + exp ⁡ ( − 1 ⁄ ( 1 − x ) ) ( 0 ≤ x ≤ 1 )

SmoothTanh(f32)

A smooth transition function using the hyperbolic tangent function (with a parameter α):
ϕ ( x ) = 1 2 ( tanh ⁡ ( α ( 2 x − 1 ) 2 x ( 1 − x ) ) + 1 ) ( 0 ≤ x ≤ 1 )

Implementations

impl BlendMethod

pub fn factor(self, ratio: f32) -> f32

Factor for blending, with ratio being the progress of transition (0 for departure, 1 for destination).

All three blending methods are symmetric with respect to (0.5, 0.5):

use BlendMethod::*;
for method in [Linear, Smooth, SmoothTanh(1.5), SmoothTanh(2.0)] {
    assert_eq!(method.factor(0.5), 0.5);
}

pub fn blend<T>(self, start: T, end: T, progress: f32) -> Twhere
T: Add<Output = T> + Mul<f32, Output = T>,

Blend two values using the specified information.

This function is implemented using multiplication and addition. For vector or matrix types with dedicated linear interpolation support, such as Vec2, Vec3, Mat2, Mat3, etc., one should prefer using BlendMethod::factor to calculate the factor manually, and use the inherent lerp method (e.g. Vec2::lerp) on those types.