Trait libre_pvz_animation::curve::AnyComponent

pub trait AnyComponent<Target: ?Sized = dyn Any> {
    // Required methods
    fn component(&self) -> &Target;
    fn component_mut(&mut self) -> &mut Target;
}

Bevy tracks changes in mutable references like Mut. We use this interface to avoid unnecessarily marking the target component as changed.

Required Methods

fn component(&self) -> &Target

Read the current value of the component.

fn component_mut(&mut self) -> &mut Target

Borrow the component as mutable and mark it dirty.